+++ CHANGELOG.md
@@ -3,6 +3,6 @@
 ## v0.0.1 - in progress
 
 ### Added
--
+- Basic schemas.
 ### Changed
 -